manejo errores en sucursal-service

+++ sucursal-service/src/main/java/com/barbershop/sucursal_service/service/SucursalService.java
@@ -1,7 +1,9 @@
 package com.barbershop.sucursal_service.service;
 
+import com.barbershop.sucursal_service.exception.ResourceNotFoundException;
 import com.barbershop.sucursal_service.model.Sucursal;
 import com.barbershop.sucursal_service.repository.SucursalRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,14 +20,25 @@ public class SucursalService {
     }
 
     public Sucursal findById(Long id) {
-        return sucursalRepository.findById(id).orElse(null);
+
+        return sucursalRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException(
+                                "Sucursal con ID " + id + " no encontrada"));
     }
 
     public Sucursal save(Sucursal sucursal) {
+
         return sucursalRepository.save(sucursal);
     }
 
     public void delete(Long id) {
+
+        if (!sucursalRepository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Sucursal con ID " + id + " no encontrada");
+        }
+
         sucursalRepository.deleteById(id);
     }
